refactor: carve out seams for controller-layer testing
Add a CandidatePanelDisplaying protocol so CandidateManager accepts an
injected panel (NSPanel is not usable from the headless swiftc test
binary). Hold the session in SessionCoordinator as LexSessionProtocol
rather than the concrete LexSession class so tests can substitute an
in-memory fake without crossing the FFI boundary. Move the
InputSource / LeximeInputSourceID helpers out of AppContext.swift into
their own file so tests pulling in ModeController do not drag in the
full AppContext bootstrap graph.
